Rename AST nodes to avoid conflicting class names
