+++ core/trino-parser/src/main/java/io/trino/sql/tree/AstVisitor.java
@@ -1097,12 +1097,12 @@ public abstract class AstVisitor<R, C>
         return visitNode(node, context);
     }
 
-    protected R visitTableArgument(TableArgument node, C context)
+    protected R visitTableArgument(TableFunctionTableArgument node, C context)
     {
         return visitNode(node, context);
     }
 
-    protected R visitDescriptorArgument(DescriptorArgument node, C context)
+    protected R visitDescriptorArgument(TableFunctionDescriptorArgument node, C context)
     {
         return visitNode(node, context);
     }
